// src/api/gen/models/ProgramVersion.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProgramVersionSummary } from "./ProgramVersionSummary";
export type ProgramVersion = ProgramVersionSummary & {
  programId?: string;
  createdAt?: string;
  updatedAt?: string;
};